+++ resources/views/horses/pedigree/index.blade.php
@@ -24,7 +24,7 @@
                                                 </p>
                                             </li>
                                             <li class="col-sm-4">
-                                                @if (! auth()->user()->isInHorseTeam($horse))
+                                                @if (auth()->check() && ! auth()->user()->isInHorseTeam($horse))
                                                     @if (Auth::user()->isFollowing($horse))
                                                         {{ Form::open(['route' => ['follows.destroy', $horse->id()], 'method' => 'DELETE']) }}
                                                         <button type="submit" class="btn btn-mint">{{ trans('copy.a.unfollow') . $horse->name() }}</button>
@@ -56,7 +56,9 @@
                                                     @include('horses.pedigree._partials._pedigree', ['family' => $horse->father()])
                                                 @else
                                                     <div class="panel-body">
-                                                        <a href="{{ route('pedigree.create', [$horse->slug(), 'type' => \EQM\Models\Pedigrees\Pedigree::FATHER]) }}" class="btn btn-mint">{{ trans('copy.a.add_father') }}</a>
+                                                        @if (auth()->check() && auth()->user()->isInHorseTeam($horse))
+                                                            <a href="{{ route('pedigree.create', [$horse->slug(), 'type' => \EQM\Models\Pedigrees\Pedigree::FATHER]) }}" class="btn btn-mint">{{ trans('copy.a.add_father') }}</a>
+                                                        @endif
                                                     </div>
                                                 @endif
                                             </div>
@@ -70,7 +72,9 @@
                                                             @include('horses.pedigree._partials._pedigree', ['family' => $horse->fathersFather()])
                                                         @else
                                                             <div class="panel-body">
-                                                                <a href="{{ route('pedigree.create', [$horse->slug(), 'type=6']) }}" class="btn btn-mint">{{ trans('copy.a.add_father_mother') }}</a>
+                                                                @if (auth()->check() && auth()->user()->isInHorseTeam($horse))
+                                                                    <a href="{{ route('pedigree.create', [$horse->slug(), 'type=6']) }}" class="btn btn-mint">{{ trans('copy.a.add_father_mother') }}</a>
+                                                                @endif
                                                             </div>
                                                         @endif
                                                     </div>
@@ -81,7 +85,9 @@
                                                             @include('horses.pedigree._partials._pedigree', ['family' => $horse->fathersMother()])
                                                         @else
                                                             <div class="panel-body">
-                                                                <a href="{{ route('pedigree.create', [$horse->slug(), 'type=6']) }}" class="btn btn-mint">{{ trans('copy.a.add_father_mother') }}</a>
+                                                                @if (auth()->check() && auth()->user()->isInHorseTeam($horse))
+                                                                    <a href="{{ route('pedigree.create', [$horse->slug(), 'type=6']) }}" class="btn btn-mint">{{ trans('copy.a.add_father_mother') }}</a>
+                                                                @endif
                                                             </div>
                                                         @endif
                                                     </div>
@@ -97,7 +103,9 @@
                                                     @include('horses.pedigree._partials._pedigree', ['family' => $horse->mother()])
                                                 @else
                                                     <div class="panel-body">
-                                                        <a href="{{ route('pedigree.create', [$horse->slug(), 'type' => \EQM\Models\Pedigrees\Pedigree::MOTHER]) }}" class="btn btn-mint">{{ trans('copy.a.add_mother') }}</a>
+                                                        @if (auth()->check() && auth()->user()->isInHorseTeam($horse))
+                                                            <a href="{{ route('pedigree.create', [$horse->slug(), 'type' => \EQM\Models\Pedigrees\Pedigree::MOTHER]) }}" class="btn btn-mint">{{ trans('copy.a.add_mother') }}</a>
+                                                        @endif
                                                     </div>
                                                 @endif
                                             </div>
@@ -111,7 +119,9 @@
                                                             @include('horses.pedigree._partials._pedigree', ['family' => $horse->mothersFather()])
                                                         @else
                                                             <div class="panel-body">
-                                                                <a href="{{ route('pedigree.create', [$horse->slug(), 'type=6']) }}" class="btn btn-mint">{{ trans('copy.a.add_father_mother') }}</a>
+                                                                @if (auth()->check() && auth()->user()->isInHorseTeam($horse))
+                                                                    <a href="{{ route('pedigree.create', [$horse->slug(), 'type=6']) }}" class="btn btn-mint">{{ trans('copy.a.add_father_mother') }}</a>
+                                                                @endif
                                                             </div>
                                                         @endif
                                                     </div>
@@ -122,7 +132,9 @@
                                                             @include('horses.pedigree._partials._pedigree', ['family' => $horse->mothersMother()])
                                                         @else
                                                             <div class="panel-body">
-                                                                <a href="{{ route('pedigree.create', [$horse->slug(), 'type=6']) }}" class="btn btn-mint">{{ trans('copy.a.add_father_mother') }}</a>
+                                                                @if (auth()->check() && auth()->user()->isInHorseTeam($horse))
+                                                                    <a href="{{ route('pedigree.create', [$horse->slug(), 'type=6']) }}" class="btn btn-mint">{{ trans('copy.a.add_father_mother') }}</a>
+                                                                @endif
                                                             </div>
                                                         @endif
                                                     </div>
@@ -173,105 +185,4 @@
             </div>
         </div>
     </div>
-    {{--@if (auth()->user()->isInHorseTeam($horse))
-        <div class="row">
-            <a href="{{ route('pedigree.create', $horse->slug()) }}" class="btn">{{ trans('copy.a.add_family') }}</a>
-        </div>
-    @endif--}}
-    {{--<div class="row pedigree">
-        <div class="col s12">
-            @if ($horse->hasFather())
-                <div class="col s3 grandparent male">
-                    @if ($family = $horse->fathersFather())
-                        @include('horses.pedigree._partials._pedigree')
-                    @else
-                        @can('create-pedigree', $horse)
-                            <a href="{{ route('pedigree.create', [$horse->slug, 'type=5']) }}" class="black-text">{{ trans('copy.a.add_father_father') }}</a>
-                        @endcan
-                    @endif
-                </div>
-                <div class="col s3 grandparent female">
-                    @if ($family = $horse->fathersMother())
-                        @include('horses.pedigree._partials._pedigree')
-                    @else
-                        @can('create-pedigree', $horse)
-                            <a href="{{ route('pedigree.create', [$horse->slug, 'type=6']) }}" class="black-text">{{ trans('copy.a.add_father_mother') }}</a>
-                        @endcan
-                    @endif
-                </div>
-            @endif
-
-            @if ($horse->hasMother())
-                <div class="col s3 grandparent male">
-                    @if ($family = $horse->mothersFather())
-                        @include('horses.pedigree._partials._pedigree')
-                    @else
-                        @can ('create-pedigree', $horse)
-                            <a href="{{ route('pedigree.create', [$horse->slug(), 'type=7']) }}" class="black-text">{{ trans('copy.a.add_mother_father') }}</a>
-                        @endcan
-                    @endif
-                </div>
-                <div class="col s3 grandparent female">
-                    @if ($family = $horse->mothersMother())
-                        @include('horses.pedigree._partials._pedigree')
-                    @else
-                        @can('create-pedigree', $horse)
-                            <a href="{{ route('pedigree.create', [$horse->slug(), 'type=8']) }}" class="black-text">{{ trans('copy.a.add_mother_mother') }}</a>
-                        @endcan
-                    @endif
-                </div>
-            @endif
-        </div>
-    </div>
-    <div class="row pedigree">
-        <div class="col s12">
-            <div class="col s6 parent male">
-                @if ($family = $horse->father())
-                    @include('horses.pedigree._partials._pedigree')
-                @else
-                    @can('create-pedigree', $horse)
-                        <a href="{{ route('pedigree.create', [$horse->slug(), 'type=1']) }}" class="black-text">{{ trans('copy.a.add_father') }}</a>
-                    @endcan
-                @endif
-            </div>
-            <div class="col s6 parent female">
-                @if ($family = $horse->mother())
-                    @include('horses.pedigree._partials._pedigree')
-                @else
-                    @can('create-pedigree', $horse)
-                        <a href="{{ route('pedigree.create', [$horse->slug(), 'type=2']) }}" class="black-text">{{ trans('copy.a.add_mother') }}</a>
-                    @endcan
-                @endif
-            </div>
-        </div>
-    </div>
-    <div class="row pedigree">
-        <div class="col s12">
-            <div class="col s12 self">
-                <a href="{{ route('horses.show', $horse->slug()) }}">
-                    <h4>{{ $horse->name() }}</h4>
-                </a>
-                <p>{{ trans('copy.p.born') . ' ' . eqm_date($horse->dateOfBirth(), 'Y') }}</p>
-                <p>{{ trans('copy.p.life_number') . ' ' . $horse->lifeNumber() ? : '-' }}</p>
-            </div>
-        </div>
-    </div>
-    <div class="row pedigree">
-        <div class="col s12">
-            <div class="col s6 offspring male">
-                {{ trans('copy.titles.sons') }}
-                <hr/>
-                @foreach ($horse->sons() as $son)
-                    {{ $son->originalHorse->name() }}
-                @endforeach
-            </div>
-            <div class="col s6 offspring female">
-                {{ trans('copy.titles.daughters') }}
-                <hr/>
-                @foreach ($horse->daughters() as $daughter)
-                    {{ $daughter->originalHorse->name() }}
-                @endforeach
-            </div>
-        </div>
-    </div>--}}
 @stop
